Fix Cycle Holding FileImageWidget Open
The gesture and overlay handlers both had lambdas, which appear to hold
"this" open, which prevented this from getting cleaned up when the
widget was removed from the tree.

Currently the level above this is always FileWidget, and FileWidget is
hard-coded to call "dispose" manually, but if this widget gets put into
another Widget that isn't special, it leaks.

This fixes this.

(cherry picked from commit da8b5418043e509275b83bd8e7b16ba6775bdefa)
Signed-off-by: Vadim Lomovtsev <jelezny@gmail.com>