Unlink Overlay After It's Closed
Previously the Overlay was a stack variable, and the only reason it
showed up was because it was leaked in the lambdas.

But making the references weak just exposes that this currently doesn't
have an obvious owner. So I made the view controller the owner, and then
in close I just clear that reference out so it can be reaped.

But, I can't free it out while in a handler on its own signal or it
crashes, so I have to wait for the next tick first.

(cherry picked from commit 6d19e8d00dd6b64d6d0256c184724c7f7424c1d9)
Signed-off-by: Vadim Lomovtsev <jelezny@gmail.com>